# api/openapi.yaml
openapi: 3.1.0
info:
  title: LingoGrade API
  version: 1.0.0
  description: Checkout, webhooks, stickers, and partner endpoints.
  contact:
    email: marco@lingograde.com

servers:
  - url: https://api.lingograde.com
    description: Production

paths:
  /v1/checkout/accessory:
    post:
      summary: Checkout for merch (cap, bracelet, pin)
      tags: [Checkout]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, product]
              properties:
                email: { type: string, format: email }
                product: { type: string, enum: [cap, bracelet, pin] }
      responses:
        "200":
          description: Stripe checkout URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkout_url: { type: string }
                  session_id: { type: string }

  /v1/checkout/kids:
    post:
      summary: Checkout for kids assessment
      tags: [Checkout]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [parent_email, child_name, age_group, package]
              properties:
                parent_email: { type: string, format: email }
                child_name: { type: string }
                age_group: { type: string, enum: ["6-8", "9-11", "12-14", "15-17"] }
                package: { type: string, enum: [quick, full, deep-dive] }
                currency: { type: string, default: eur }
                guardian_name: { type: string }
      responses:
        "200":
          description: Stripe checkout URL

  /v1/checkout/mega-bundle:
    post:
      summary: Checkout for Mega Bundle
      tags: [Checkout]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                currency: { type: string, default: eur }
      responses:
        "200":
          description: Stripe checkout URL

  /v1/stripe/webhook:
    post:
      summary: Stripe webhook handler
      tags: [Webhooks]
      description: Handles payment_intent.succeeded and checkout.session.completed events.
      responses:
        "200": { description: OK }

  /v1/stickers/map:
    get:
      summary: Public sticker placement map data
      tags: [Stickers]
      responses:
        "200":
          description: Sticker placement locations and stats

  /v1/stickers/verify:
    post:
      summary: Submit sticker selfie for verification
      tags: [Stickers]
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [sticker_uuid, latitude, longitude, selfie]
              properties:
                sticker_uuid: { type: string }
                latitude: { type: number }
                longitude: { type: number }
                city: { type: string }
                country: { type: string }
                selfie: { type: string, format: binary }
      responses:
        "201":
          description: Verification submitted

  /api/partner-apply:
    post:
      summary: Submit partner application
      tags: [Partners]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                phone: { type: string }
                institution: { type: string }
                country: { type: string }
                languages: { type: string }
                referral_source: { type: string }
      responses:
        "200": { description: Application received }

  /v1/config:
    get:
      summary: Public frontend config (Stripe publishable key)
      tags: [Config]
      responses:
        "200":
          description: Config object

  /health:
    get:
      summary: Health check
      tags: [Config]
      responses:
        "200":
          description: Service status

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer